// src/api/types.ts
export type User={
    auth0Id: string;
    email: string;
}
export type UpdateUser ={
    name: string;
    address: string;
    country: string;
    city: string;
}

export type BackEndUser ={
    _id: string;
    email: string;
    name: string;
    address: string;
    country: string;
    city: string;
}